Connect & Asset Minification Changes
Version 1 of Twilio Connect support. Everything works, but needs a
bit of polish. Added welcome controller. Abstracted steps concept
for later rebuilding of install and upgrade to use the generic
library.

Adding Minify library for minification of assets on the server
to remove the need to recompile assets after changes and to better
version the assets on delivery with consistent url version scheme.