Upgrade instructions! & asset url fixes
Upgrade instructions finally exist. There's no accounting for the
quality of writing, but its there ;)

Also a fix for asset urls on added scripts;

Also a tweak for the installer that notifies the user in the event
that the .htaccess file is missing. Adding a `htaccess_dist` file
to make restoring the .htaccess file easier.