Fixing invite code generation & update
Invite code generation was making characters that wanted to be
encoded in a url and hence causing mayhem during password resets.
Updated to base64 encode the invite code to make it url safe.

Moved the password & dial timeout updates to a new schema version
so as not to conflict with the 64 schema version introduced with
1.1.3