Moving version number to file & other...
Moved the version number to be in a static file instead of being
pulled from the database. This allows us to inspect the version
during install to properly set API request user agent.

Updates to items that used to pull the version directly from the
database instead of using `OpenVBX::version()`

Other minor fixes, updates, and formatting adjustments