safety checks
Explicitly setting the timeLimit on the Dial when starting a
conference due to submitted bug report that it is defaulting
to 1 hour. Setting as a precaution.

Disabling upgrade check to GitHub tags if `open_basedir` or php
`safe_mode` are set as they'll cause `CURLOPT_FOLLOWLOCATION` to
not set and also cause all the rest of the curlopt settings to
not set.