Sidebar Menu Navigation: Bug Fix
Bug: The code only hightlights the menu as being selected if the pages are under "Messages, Setup and Admin" sections.  If the pages are from a plugin they will never get marked as being selected because the $section and the $link never match.

Fix: The easiest fix was to add a forward slash into the if statement when comparing the current section to the current link.  That was the only differece I could find and appeared to be the only reason why the code would never highlight the menu the user selected which clicking a plugin page.